Move setupApp
